+++ include/namz.h
@@ -45,7 +45,7 @@
 #define NAMZ_VERSION_MAJOR 3
 #define NAMZ_VERSION_MINOR 3
 #define NAMZ_VERSION_PATCH 0
-#define NAMZ_VERSION "3.3.0"
+#define NAMZ_VERSION "4.0.0"
 
 #include <cstddef>
 #include <cstdint>